Split ingest_events and auth out from server
